+++ app/Http/Controllers/ProfileController.php
@@ -8,6 +8,10 @@ use App\Http\Resources\UserResource;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use Endroid\QrCode\QrCode;
+use Endroid\QrCode\Writer\PngWriter;
+use PragmaRX\Google2FA\Google2FA;
+use App\Http\Controllers\redirect;
 
 class ProfileController extends Controller
 {
@@ -43,10 +47,31 @@ class ProfileController extends Controller
         return view('profile.change-password');
     }
 
-    public function showTwoFactor()
+    public function active2FA()
     {
-        $activeTab = 3;
-        return view('profile.two-factor');
+        $user = Auth::user();
+        $user->refresh();
+
+        $google2fa = new Google2FA();
+
+        $qrCodeUrl = $google2fa->getQRCodeUrl(
+            config('app.name'),
+            $user->email,
+            $user->two_factor_secret
+        );
+
+        $qrCode = new QrCode($qrCodeUrl);
+        $writer = new PngWriter();
+
+        $qrImage = $writer->write($qrCode);
+        $qrImageBase64 = base64_encode($qrImage->getString());
+
+        $recoveryCodes = null;
+        if (isset($user->two_factor_recovery_codes)) {
+            $recoveryCodes = json_decode(decrypt($user->two_factor_recovery_codes), true);
+        }
+
+        return view('auth.two-factor-settings', compact('qrImageBase64', 'recoveryCodes'));
     }
 
     public function updateUserData(UpdateUserRequest $request)
